source/requirements.txt
discord.py
python-gitlab